add Get list product by category, include paging, sorting

+++ src/main/java/com/ecommerceserver/model/Product.java
@@ -15,7 +15,7 @@ public class Product {
   String id;
 
   String title;
-  BigDecimal price;
+  Long price;
   String description;
   int stock;
   String imageUrl;
